# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/client-access": {
      "post": {
        "operationId": "ClientAccess_grant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantAccessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantAccessResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ClientAccess"
        ]
      }
    },
    "/client-access/{id}": {
      "delete": {
        "operationId": "ClientAccess_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ClientAccess"
        ]
      }
    },
    "/client-access/updates/{projectId}": {
      "post": {
        "operationId": "ClientAccess_publishUpdate",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishUpdateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUpdate"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ClientAccess"
        ]
      }
    },
    "/p/{token}": {
      "get": {
        "operationId": "ClientPortal_view",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientProjectViewDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ClientPortal"
        ]
      }
    },
    "/p/{token}/offers": {
      "post": {
        "operationId": "ClientPortal_requestOffer",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOfferDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ClientPortal"
        ]
      }
    }
  },
  "info": {
    "title": "Snapline API",
    "description": "Gestión para contratistas. Ver docs/product/vision.md",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "billingAddress": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "source": {
            "nullable": true,
            "enum": [
              "REFERRAL",
              "WEB",
              "SOCIAL",
              "REPEAT",
              "OTHER"
            ],
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "photoReleaseGrantedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "photoReleaseDocumentId": {
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "firstName",
          "lastName",
          "companyName",
          "email",
          "phone",
          "billingAddress",
          "source",
          "notes",
          "photoReleaseGrantedAt",
          "photoReleaseDocumentId",
          "deletedAt",
          "companyId",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Site": {
        "type": "object",
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "customerId": {
            "type": "string"
          },
          "address": {
            "type": "object",
            "additionalProperties": true
          },
          "lat": {
            "type": "number",
            "nullable": true
          },
          "lng": {
            "type": "number",
            "nullable": true
          },
          "geofenceRadiusM": {
            "type": "number",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "customer",
          "customerId",
          "address",
          "lat",
          "lng",
          "geofenceRadiusM",
          "deletedAt",
          "companyId",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "customerId": {
            "type": "string"
          },
          "site": {
            "$ref": "#/components/schemas/Site"
          },
          "siteId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "serviceType": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "COMPLETED",
              "LEAD",
              "ESTIMATED",
              "SCHEDULED",
              "IN_PROGRESS",
              "ON_HOLD",
              "CANCELLED"
            ],
            "type": "string"
          },
          "clientVisibilityMode": {
            "enum": [
              "STAGES",
              "PROGRESS"
            ],
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "nullable": true
          },
          "targetEndDate": {
            "type": "string",
            "nullable": true
          },
          "actualEndDate": {
            "type": "string",
            "nullable": true
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "customer",
          "customerId",
          "site",
          "siteId",
          "name",
          "description",
          "serviceType",
          "status",
          "clientVisibilityMode",
          "startDate",
          "targetEndDate",
          "actualEndDate",
          "publishedAt",
          "deletedAt",
          "companyId",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "AppUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "enum": [
              "en",
              "es"
            ]
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "phone",
          "name",
          "locale",
          "deletedAt",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Membership": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AppUser"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "FOREMAN",
              "WORKER",
              "ACCOUNTANT"
            ]
          },
          "payRateCents": {
            "type": "number",
            "nullable": true
          },
          "employmentType": {
            "type": "string",
            "nullable": true,
            "enum": [
              "W2",
              "CONTRACTOR_1099"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "invitedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "user",
          "userId",
          "role",
          "payRateCents",
          "employmentType",
          "status",
          "invitedAt",
          "joinedAt",
          "deletedAt",
          "companyId",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "GrantAccessDto": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "description": "Null da acceso a todos los proyectos del cliente.",
            "format": "uuid"
          },
          "expiresInDays": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          }
        },
        "required": [
          "customerId"
        ]
      },
      "GrantAccessResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "description": "Se muestra una sola vez: en la base solo queda su hash."
          },
          "url": {
            "type": "string",
            "description": "Link listo para mandar por SMS o email."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "token",
          "url",
          "expiresAt"
        ]
      },
      "PublishUpdateDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          },
          "assetIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "body"
        ]
      },
      "ProjectUpdate": {
        "type": "object",
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/Membership"
          },
          "authorMembershipId": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "INTERNAL",
              "CLIENT",
              "PUBLIC"
            ]
          },
          "approvedBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/Membership"
              }
            ]
          },
          "approvedByMembershipId": {
            "type": "string",
            "nullable": true
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "companyId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "project",
          "projectId",
          "author",
          "authorMembershipId",
          "body",
          "visibility",
          "approvedBy",
          "approvedByMembershipId",
          "publishedAt",
          "deletedAt",
          "companyId",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "ClientPhotoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "url",
          "capturedAt"
        ]
      },
      "ClientUpdateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientPhotoDto"
            }
          }
        },
        "required": [
          "id",
          "body",
          "publishedAt",
          "photos"
        ]
      },
      "ClientOfferDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "pitch": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "pitch"
        ]
      },
      "ClientProjectViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "stage": {
            "enum": [
              "INICIO",
              "EN_PROCESO",
              "FINALIZADO"
            ],
            "type": "string"
          },
          "visibilityMode": {
            "type": "string",
            "enum": [
              "STAGES",
              "PROGRESS"
            ],
            "description": "En STAGES el cliente solo ve la etapa; updates y fotos vienen vacíos."
          },
          "updates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientUpdateDto"
            }
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientPhotoDto"
            }
          },
          "offers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientOfferDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "stage",
          "visibilityMode",
          "updates",
          "photos",
          "offers"
        ]
      },
      "RequestOfferDto": {
        "type": "object",
        "properties": {
          "offerId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "offerId"
        ]
      },
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "code": {
            "enum": [
              "BAD_REQUEST",
              "VALIDATION_FAILED",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "SERVICE_UNAVAILABLE",
              "INTERNAL_ERROR",
              "INVALID_CREDENTIALS",
              "TOKEN_MISSING",
              "TOKEN_INVALID",
              "MEMBERSHIP_INACTIVE",
              "PERMISSION_NOT_DECLARED",
              "PERMISSION_DENIED",
              "TIME_ENTRY_ALREADY_OPEN",
              "TIME_ENTRY_ALREADY_CLOSED",
              "CANNOT_APPROVE_OWN_HOURS",
              "PAY_RATE_MISSING",
              "PHOTO_RELEASE_REQUIRED",
              "EXIF_NOT_STRIPPED",
              "UPLOAD_NOT_READY",
              "ASSET_NOT_PUBLIC",
              "ALREADY_PUBLISHED",
              "ESTIMATE_ALREADY_SENT",
              "ESTIMATE_NOT_ACCEPTED",
              "ESTIMATE_ALREADY_INVOICED",
              "INVOICE_NOT_SENT",
              "INVOICE_VOIDED",
              "PAYMENT_EXCEEDS_BALANCE",
              "STORAGE_NOT_CONFIGURED"
            ],
            "type": "string",
            "description": "Código estable. No se traduce: es contra lo que ramifica el cliente."
          },
          "message": {
            "type": "string",
            "description": "Texto para mostrar. Esto sí se traduce."
          },
          "details": {
            "description": "Vacío cuando no aplica; nunca ausente.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "path",
          "timestamp"
        ]
      }
    }
  }
}